cookiecutter-django-template/{{cookiecutter.project_dir}}/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
requests

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-dbapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-sqlite3
opentelemetry-instrumentation-threading
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib3